Move android app to separate package
