+++ app/src/main/kotlin/io/github/dector/quotes/android/QuotesModule.kt
@@ -1,14 +1,14 @@
-package io.github.dector.quotes.qoutes
+package io.github.dector.quotes.android
 
 import android.content.Context
 import android.view.LayoutInflater
 import dagger.Module
 import dagger.Provides
+import io.github.dector.qoutes.storage.MockQuotesStorage
 import io.github.dector.quotes.R
+import io.github.dector.quotes.android.presentation.view.QuotesView
 import io.github.dector.quotes.domain.Quote
 import io.github.dector.quotes.presentation.presenter.QuotesPresenter
-import io.github.dector.quotes.qoutes.presentation.view.QuotesView
-import io.github.dector.qoutes.storage.MockQuotesStorage
 import io.github.dector.quotes.storage.IStorage
 import io.github.dector.quotes.usecases.IQuotesUseCase
 import io.github.dector.quotes.usecases.QuotesUseCase
