+++ app/src/main/kotlin/io/github/dector/quotes/android/common/Rx.kt
@@ -1,4 +1,4 @@
-package io.github.dector.quotes.common
+package io.github.dector.quotes.android.common
 
 import rx.subjects.BehaviorSubject
